--- src/main/java/com/example/capstonedesigin20252/CapstoneDesigin20252Application.java
@@ -1,13 +1,0 @@
-package com.example.capstonedesigin20252;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CapstoneDesigin20252Application {
-
-  public static void main(String[] args) {
-    SpringApplication.run(CapstoneDesigin20252Application.class, args);
-  }
-
-}
