--- src/main/java/com/example/capstonedesigin20252/common/exception/ErrorCode.java
@@ -1,10 +1,0 @@
-package com.example.capstonedesigin20252.common.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ErrorCode {
-
-  String getMessage();
-  String getCode();
-  HttpStatus getStatus();
-}
