--- src/main/java/com/example/capstonedesigin20252/common/exception/ErrorResponseDto.java
@@ -1,6 +1,0 @@
-package com.example.capstonedesigin20252.common.exception;
-
-public record ErrorResponseDto(
-    String message, String code, int status
-) {
-}
