--- src/main/java/com/example/capstonedesigin20252/paymentLog/service/PaymentLogService.java
@@ -1,7 +1,0 @@
-package com.example.capstonedesigin20252.service;
-
-import com.example.capstonedesigin20252.dto.PaymentRequestDto;
-
-public interface PaymentLogService {
-  void savePaymentLog(PaymentRequestDto paymentRequestDto);
-}
