--- src/main/java/com/example/capstonedesigin20252/repository/PaymentLogRepository.java
@@ -1,8 +1,0 @@
-package com.example.capstonedesigin20252.repository;
-
-import com.example.capstonedesigin20252.domain.PaymentLog;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PaymentLogRepository extends JpaRepository<PaymentLog, Long> {
-
-}
