--- src/main/java/com/example/capstonedesigin20252/service/PaymentLogServiceImpl.java
@@ -1,22 +1,0 @@
-package com.example.capstonedesigin20252.service;
-
-import com.example.capstonedesigin20252.domain.PaymentLog;
-import com.example.capstonedesigin20252.dto.PaymentRequestDto;
-import com.example.capstonedesigin20252.repository.PaymentLogRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class PaymentLogServiceImpl implements PaymentLogService {
-
-  private final PaymentLogRepository paymentLogRepository;
-
-  @Transactional
-  public void savePaymentLog(PaymentRequestDto paymentRequestDto){
-    PaymentLog paymentLog = paymentRequestDto.toEntity();
-    paymentLogRepository.save(paymentLog);
-  }
-}
