--- src/main/java/com/example/capstonedesigin20252/user/domain/LoginType.java
@@ -1,5 +1,0 @@
-package com.example.capstonedesigin20252.user.domain;
-
-public enum LoginType {
-  GENERAL, KAKAO
-}
