--- src/main/java/com/example/capstonedesigin20252/user/dto/LoginRequestDto.java
@@ -1,15 +1,0 @@
-package com.example.capstonedesigin20252.user.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginRequestDto(
-
-    @NotBlank(message = "이메일은 필수입니다.")
-    @Email(message = "올바른 이메일 형식이 아닙니다")
-    String email,
-
-    @NotBlank(message = "비밀번호는 필수입니다.")
-    String password
-) {
-}
