+++ src/main/java/com/example/capstonedesign20252/common/exception/BaseException.java
@@ -1,0 +1,9 @@
+package com.example.capstonedesign20252.common.exception;
+
+public abstract class BaseException extends RuntimeException {
+
+  public abstract String getMessage();
+  public abstract String getCode();
+  public abstract int getStatus();
+
+}
