+++ src/main/java/com/example/capstonedesign20252/common/exception/ErrorCode.java
@@ -1,0 +1,10 @@
+package com.example.capstonedesign20252.common.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorCode {
+
+  String getMessage();
+  String getCode();
+  HttpStatus getStatus();
+}
