+++ src/main/java/com/example/capstonedesign20252/common/exception/ErrorResponseDto.java
@@ -1,0 +1,6 @@
+package com.example.capstonedesign20252.common.exception;
+
+public record ErrorResponseDto(
+    String message, String code, int status
+) {
+}
