+++ src/main/java/com/example/capstonedesign20252/paymentLog/dto/PaymentRequestDto.java
@@ -1,0 +1,17 @@
+package com.example.capstonedesign20252.dto;
+
+import com.example.capstonedesign20252.domain.PaymentLog;
+
+public record PaymentRequestDto(
+    String name,
+    Integer amount,
+    String targetAccount
+) {
+  public PaymentLog toEntity(){
+    return PaymentLog.builder()
+        .name(this.name)
+        .amount(this.amount)
+        .targetAccount(this.targetAccount)
+        .build();
+  }
+}
