+++ src/main/java/com/example/capstonedesign20252/paymentLog/repository/PaymentLogRepository.java
@@ -1,0 +1,8 @@
+package com.example.capstonedesign20252.repository;
+
+import com.example.capstonedesign20252.domain.PaymentLog;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PaymentLogRepository extends JpaRepository<PaymentLog, Long> {
+
+}
