+++ src/main/java/com/example/capstonedesign20252/paymentLog/service/PaymentLogService.java
@@ -1,0 +1,7 @@
+package com.example.capstonedesign20252.service;
+
+import com.example.capstonedesign20252.dto.PaymentRequestDto;
+
+public interface PaymentLogService {
+  void savePaymentLog(PaymentRequestDto paymentRequestDto);
+}
