+++ src/main/java/com/example/capstonedesign20252/user/domain/LoginType.java
@@ -1,0 +1,5 @@
+package com.example.capstonedesign20252.user.domain;
+
+public enum LoginType {
+  GENERAL, KAKAO
+}
