+++ src/main/java/com/example/capstonedesign20252/user/dto/SignUpResponseDto.java
@@ -1,0 +1,22 @@
+package com.example.capstonedesign20252.user.dto;
+
+import com.example.capstonedesign20252.user.domain.User;
+import java.time.LocalDateTime;
+import lombok.Builder;
+
+@Builder
+public record SignUpResponseDto(
+    Long userId,
+    String name,
+    String email,
+    LocalDateTime createdAt
+) {
+  public static SignUpResponseDto from(User user){
+    return SignUpResponseDto.builder()
+      .userId(user.getId())
+      .name(user.getName())
+      .email(user.getEmail())
+      .createdAt(user.getCreatedAt())
+      .build();
+  }
+}
