+++ src/main/java/com/example/capstonedesign20252/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.example.capstonedesign20252.user.repository;
+
+import com.example.capstonedesign20252.user.domain.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+  Optional<User> findByEmail(String email);
+  boolean existsByEmail(String email);
+
+}
