--- src/test/java/com/example/capstonedesigin20252/CapstoneDesigin20252ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.capstonedesigin20252;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CapstoneDesigin20252ApplicationTests {
-
-  @Test
-  void contextLoads() {
-  }
-
-}
