+++ src/test/java/com/example/capstonedesign20252/CapstoneDesigin20252ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.capstonedesign20252;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CapstoneDesign20252ApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
